adding the filter responsibilities to the Filter component
